Falta poner comentarios a las fucnniones

+++ matrix.h
@@ -6,7 +6,7 @@
 /*   By: mvelazqu <mvelazqu@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/19 18:43:44 by mvelazqu          #+#    #+#             */
-/*   Updated: 2024/03/23 17:44:55 by mvelazqu         ###   ########.fr       */
+/*   Updated: 2024/03/24 19:03:05 by mvelazqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,11 +27,11 @@ t_matrix	create_matrix(t_vector vec_x, t_vector vec_y, t_vector vec_z);
 t_matrix	create_rotation_x(float angle);
 t_matrix	create_rotation_y(float angle);
 t_matrix	create_rotation_z(float angle);
-t_matrix	create_scale(int scale, int percentage);
+t_matrix	create_scale(float scale, int percentage);
 t_vector	vector_x_matrix(t_vector vec, t_matrix mat);
 void		transform_object(t_point *object, t_matrix matrix);
 void		move_object(t_point *object, int distance, int direction);
-void		isometric_object(t_point *point, int scale);
-void		paralel_object(t_point *point, int scale);
+void		isometric_object(t_point *point, float scale);
+void		paralel_object(t_point *point, float scale);
 
 #endif
